Merging to keep up with changes on master.

+++ logger/fxui/src/main/java/logger/fxui/AppController.java
@@ -19,18 +19,18 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import logger.core.Building;
+import logger.core.Filter;
 import logger.core.Validation;
 import logger.core.Visit;
 import logger.fxui.utils.LocalVisitLogDataAccess;
 import logger.fxui.utils.RemoteVisitLogDataAccess;
 import logger.fxui.utils.VisitLogDataAccess;
-import logger.fxui.utils.VisitLogFilter;
 import logger.json.BuildingReader;
 
 public class AppController {
 
   // true for remote storage, false for local storage
-  private final VisitLogDataAccess visitLogDataAccess = isRemoteStorage(true);
+  private final VisitLogDataAccess visitLogDataAccess = isRemoteStorage(false);
 
   // Registration
   @FXML
@@ -139,6 +139,9 @@ public class AppController {
   @FXML
   private void deleteVisit() {
     ObservableList<Visit> deleteList = tableView.getSelectionModel().getSelectedItems();
+    if (deleteList.isEmpty()) {
+      return;
+    }
     Visit deleteVisit = deleteList.get(0);
     visitLogDataAccess.deleteVisit(deleteVisit.getId());
     updateTable();
@@ -206,11 +209,11 @@ public class AppController {
     logToDate.setVisible(searchKey.equals("Date"));
 
     result = switch (searchKey) {
-      case "Name" -> VisitLogFilter.filterByName(searchInput, allVisits);
-      case "Phone" -> VisitLogFilter.filterByPhone(searchInput, allVisits);
-      case "Building" -> VisitLogFilter.filterByBuilding(searchInput, allVisits);
-      case "Room" -> VisitLogFilter.filterByRoom(searchInput, allVisits);
-      case "Date" -> VisitLogFilter
+      case "Name" -> Filter.filterByName(searchInput, allVisits);
+      case "Phone" -> Filter.filterByPhone(searchInput, allVisits);
+      case "Building" -> Filter.filterByBuilding(searchInput, allVisits);
+      case "Room" -> Filter.filterByRoom(searchInput, allVisits);
+      case "Date" -> Filter
           .filterByDate(allVisits, logFromDate.getValue(), logToDate.getValue());
       default -> allVisits;
     };
@@ -239,7 +242,7 @@ public class AppController {
    * Make uri for endpoint.
    *
    * @param uri for endpoint
-   * @return
+   * @return a valid URI for the endpoint
    */
   private URI uriSetup(String uri) {
     URI newUri = null;
@@ -255,7 +258,7 @@ public class AppController {
    * Sets local or remote storage.
    *
    * @param isRemote either true or false
-   * @return
+   * @return an instance of remote storage if true, an instance of local storage if false
    */
   private VisitLogDataAccess isRemoteStorage(boolean isRemote) {
     if (isRemote) {
@@ -341,7 +344,8 @@ public class AppController {
    */
   private void setUpBuildings() {
     try {
-      List<Building> buildings = BuildingReader.readBuildings();
+      List<Building> buildings = BuildingReader
+          .readBuildings(getClass().getResource("buildings.json"));
       dropdownBuilding.getItems().addAll(buildings);
     } catch (IOException e) {
       System.out.println("Couldn't fetch any buildings");
